+++ Avoavo/app/src/main/java/com/yeahoohunters/avoavo/view/HomeFifthFragment.kt
@@ -1,0 +1,22 @@
+package com.yeahoohunters.avoavo.view
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.yeahoohunters.avoavo.R
+
+class HomeFifthFragment : Fragment() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home_fifth, container, false)
+    }
+}
