+++ Avoavo/app/src/main/res/layout/fragment_home_fifth.xml
@@ -4,4 +4,4 @@
     android:id="@+id/frameLayout6"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFifthFragment" />
+    tools:context=".view.HomeFifthFragment" />
