+++ Avoavo/app/src/main/res/layout/fragment_home_first.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFloor1Fragment" >
+    tools:context=".view.HomeFirstFragment" >
 
     <ImageView
         android:id="@+id/fun_1F"
