+++ Avoavo/app/src/main/res/layout/fragment_home_fourth.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout5"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFourthFragment" >
+    tools:context=".view.HomeFourthFragment" >
 
     <ImageView
         android:id="@+id/fun_4F"
