+++ Avoavo/app/src/main/res/layout/fragment_home_second.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFloor2Fragment" >
+    tools:context=".view.HomeSecondFragment" >
 
     <ImageView
         android:id="@+id/fun_2F"
