+++ Avoavo/app/src/main/res/layout/fragment_home_third.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout4"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFloor3Fragment" >
+    tools:context=".view.HomeFloor3Fragment" >
 
     <ImageView
         android:id="@+id/fun_3F"
